Add Vagrant and Docker config
I created a Vagrant image for running the GitLab CI/CD and GitHub
actions. The image includes Docker, act and GitLab Runner. Additionally
I added a Dockerfile that describes the image we use in the lecture. In
contrast to previous lectures this image is based on Alpine Linux which
creates much smaller images.