Lecture on automation (GitHub Actions, GitLab CI) (#78)
* Initial file layout for automation lecture

The automation lecture focuses on the automation of workflows using
GitHub Actions and GitLab CI (github_actions_gitlab_ci*.md).
There will be a short introduction on automation
(automation_slides.md).

* Work on automation chapter

* More additions to automation chapter

* Additional work on automation lecture

* Add Vagrant and Docker config

I created a Vagrant image for running the GitLab CI/CD and GitHub
actions. The image includes Docker, act and GitLab Runner. Additionally
I added a Dockerfile that describes the image we use in the lecture. In
contrast to previous lectures this image is based on Alpine Linux which
creates much smaller images.

* Update slides and demo notes

* Polishing of slides

* More polishing of lecture content

* More polishing

* Update timetable

* Fixed automation slides link

* Update 05_testing_and_ci/automation_slides.md

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@gmail.com>

* Update 05_testing_and_ci/examples/automation/bootstrap.sh

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@gmail.com>

* Update 05_testing_and_ci/automation_slides.md

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@gmail.com>

* Fix formulation in bootstrapping.sh

* Integrate feedback

* Incorporate further feedback

* More polishing of the lecture

* Fix formatting of GitLab yaml file

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@gmail.com>